refactor: InpuBar as subview instead of accessoryView
Cleanup of commented and not used code
Moved some internal properties to state object
Removed iOS 13 available checks
