+++ Example/Sources/Views/CameraInputBarAccessoryView.swift
@@ -71,13 +71,7 @@ extension CameraInputBarAccessoryViewDelegate {
         return InputBarButtonItem()
             .configure {
                 $0.spacing = .fixed(10)
-                   
-                if #available(iOS 13.0, *) {
-                    $0.image = UIImage(systemName: "camera.fill")?.withRenderingMode(.alwaysTemplate)
-                } else {
-                    $0.image = UIImage(named: named)?.withRenderingMode(.alwaysTemplate)
-                }
-                
+                $0.image = UIImage(systemName: "camera.fill")?.withRenderingMode(.alwaysTemplate)
                 $0.setSize(CGSize(width: 30, height: 30), animated: false)
             }.onSelected {
                 $0.tintColor = .systemBlue
