+++ Example/Sources/Views/SwiftUI/MessagesView.swift
@@ -35,7 +35,6 @@ final class MessageSwiftUIVC: MessagesViewController {
     }
 }
 
-@available(iOS 13.0, *)
 struct MessagesView: UIViewControllerRepresentable {
     
     @State var initialized = false
@@ -89,7 +88,6 @@ struct MessagesView: UIViewControllerRepresentable {
 
 }
 
-@available(iOS 13.0, *)
 extension MessagesView.Coordinator: MessagesDataSource {
     func currentSender() -> SenderType {
         return SampleData.shared.currentSender
@@ -122,7 +120,6 @@ extension MessagesView.Coordinator: MessagesDataSource {
     }
 }
 
-@available(iOS 13.0, *)
 extension MessagesView.Coordinator: InputBarAccessoryViewDelegate {
     func inputBar(_ inputBar: InputBarAccessoryView, didPressSendButtonWith text: String) {
         let message = MockMessage(text: text, user: SampleData.shared.currentSender, messageId: UUID().uuidString, date: Date())
@@ -131,7 +128,6 @@ extension MessagesView.Coordinator: InputBarAccessoryViewDelegate {
     }
 }
 
-@available(iOS 13.0, *)
 extension MessagesView.Coordinator: MessagesLayoutDelegate, MessagesDisplayDelegate {
     func configureAvatarView(_ avatarView: AvatarView, for message: MessageType, at indexPath: IndexPath, in messagesCollectionView: MessagesCollectionView) {
         let avatar = SampleData.shared.getAvatarFor(sender: message.sender)
