+++ Sources/Controllers/MessagesViewController+Keyboard.swift
@@ -30,108 +30,74 @@ import InputBarAccessoryView
 internal extension MessagesViewController {
 
     // MARK: - Register Observers
-
+    
     func addKeyboardObservers() {
-        if #available(iOS 13.0, *) {
-            NotificationCenter.default
-                .publisher(for: UITextView.textDidBeginEditingNotification)
-                .subscribe(on: DispatchQueue.global())
-                .receive(on: DispatchQueue.main)
-                .sink { [weak self] notification in
-                    self?.handleTextViewDidBeginEditing(notification)
-                }
-                .store(in: &disposeBag)
-
-            Publishers.MergeMany(
-                NotificationCenter.default.publisher(for: UIResponder.keyboardWillHideNotification),
-                NotificationCenter.default.publisher(for: UIResponder.keyboardDidHideNotification),
-                NotificationCenter.default.publisher(for: UIResponder.keyboardWillShowNotification),
-                NotificationCenter.default.publisher(for: UIResponder.keyboardDidShowNotification)
-            )
+        keyboardManager.bind(inputAccessoryView: inputContainerView)
+        keyboardManager.bind(to: messagesCollectionView)
+        
+        NotificationCenter.default
+            .publisher(for: UITextView.textDidBeginEditingNotification)
             .subscribe(on: DispatchQueue.global())
-            .sink(receiveValue: { [weak self] _ in
-                self?.updateMessageCollectionViewBottomInset()
-            })
+            .receive(on: DispatchQueue.main)
+            .sink { [weak self] notification in
+                self?.handleTextViewDidBeginEditing(notification)
+            }
             .store(in: &disposeBag)
-        }
+        
+        Publishers.MergeMany(
+            NotificationCenter.default.publisher(for: UIResponder.keyboardWillHideNotification),
+            NotificationCenter.default.publisher(for: UIResponder.keyboardDidHideNotification),
+            NotificationCenter.default.publisher(for: UIResponder.keyboardWillShowNotification),
+            NotificationCenter.default.publisher(for: UIResponder.keyboardDidShowNotification)
+        )
+        .subscribe(on: DispatchQueue.global())
+        .receive(on: DispatchQueue.main)
+        .sink(receiveValue: { [weak self] _ in
+            self?.updateMessageCollectionViewBottomInset()
+        })
+        .store(in: &disposeBag)
     }
 
-    // MARK: - Private methods
+    // MARK: - Updating insets
 
-    private func updateMessageCollectionViewBottomInset() {
-
+    /// Updates bottom messagesCollectionView inset based on the position of inputContainerView
+    func updateMessageCollectionViewBottomInset() {
         /// This is important to skip notifications from child modal controllers in iOS >= 13.0
-        guard
-            viewIsVisible,
-            self.presentedViewController == nil
-        else {
-            return
-        }
-
+        guard self.presentedViewController == nil else { return }
         let collectionViewHeight = messagesCollectionView.frame.height
         let newBottomInset = collectionViewHeight - (inputContainerView.frame.minY - additionalBottomInset) - automaticallyAddedBottomInset
         let differenceOfBottomInset = newBottomInset - messageCollectionViewBottomInset
 
         UIView.performWithoutAnimation {
-            if differenceOfBottomInset != 0 {
-                messageCollectionViewBottomInset = newBottomInset
-            }
+            guard differenceOfBottomInset != 0 else { return }
+            messagesCollectionView.contentInset.bottom = max(0, newBottomInset)
+            messagesCollectionView.verticalScrollIndicatorInsets.bottom = newBottomInset
         }
 
         if maintainPositionOnKeyboardFrameChanged && differenceOfBottomInset != 0 {
             let contentOffset = CGPoint(x: messagesCollectionView.contentOffset.x, y: messagesCollectionView.contentOffset.y + differenceOfBottomInset)
             // Changing contentOffset to bigger number than the contentSize will result in a jump of content
             // https://github.com/MessageKit/MessageKit/issues/1486
-            guard contentOffset.y <= messagesCollectionView.contentSize.height else {
-                return
-            }
+            guard contentOffset.y <= messagesCollectionView.contentSize.height else { return }
             messagesCollectionView.setContentOffset(contentOffset, animated: false)
         }
     }
 
+    // MARK: - Private methods
+
     private func handleTextViewDidBeginEditing(_ notification: Notification) {
-        if scrollsToLastItemOnKeyboardBeginsEditing || scrollsToLastItemOnKeyboardBeginsEditing {
-            guard
-                let inputTextView = notification.object as? InputTextView,
-                inputTextView === messageInputBar.inputTextView
-            else {
-                return
-            }
-            if scrollsToLastItemOnKeyboardBeginsEditing {
-                messagesCollectionView.scrollToLastItem()
-            } else {
-                messagesCollectionView.scrollToLastItem(animated: true)
-            }
+        guard scrollsToLastItemOnKeyboardBeginsEditing || scrollsToLastItemOnKeyboardBeginsEditing else { return }
+        guard
+            let inputTextView = notification.object as? InputTextView,
+            inputTextView === messageInputBar.inputTextView
+        else {
+            return
         }
-    }
-
-    // MARK: - Inset Computation
-
-    private func requiredScrollViewBottomInset(for inputViewFrame: CGRect) -> CGFloat {
-        // we only need to adjust for the part of the keyboard that covers (i.e. intersects) our collection view;
-        // see https://developer.apple.com/videos/play/wwdc2017/242/ for more details
-        let intersection = messagesCollectionView.frame.intersection(inputViewFrame)
-        if intersection.isNull || (messagesCollectionView.frame.maxY - intersection.maxY) > 0.001 {
-            // The keyboard is hidden, is a hardware one, or is undocked and does not cover the bottom of the collection view.
-            // Note: intersection.maxY may be less than messagesCollectionView.frame.maxY when dealing with undocked keyboards.
-            return max(0, additionalBottomInset - automaticallyAddedBottomInset)
+        if scrollsToLastItemOnKeyboardBeginsEditing {
+            messagesCollectionView.scrollToLastItem()
+        } else {
+            messagesCollectionView.scrollToLastItem(animated: true)
         }
-
-        let heightOfSpaceUnderInputContainer = max(0, messagesCollectionView.frame.height - intersection.maxY)
-        return heightOfSpaceUnderInputContainer
-//        if intersection.isNull || (messagesCollectionView.frame.maxY - intersection.maxY) > 0.001 {
-//            // The keyboard is hidden, is a hardware one, or is undocked and does not cover the bottom of the collection view.
-//            // Note: intersection.maxY may be less than messagesCollectionView.frame.maxY when dealing with undocked keyboards.
-//            return max(0, additionalBottomInset - automaticallyAddedBottomInset)
-//        } else {
-//            return max(0, intersection.height + additionalBottomInset - automaticallyAddedBottomInset)
-//        }
-    }
-
-    func requiredInitialScrollViewBottomInset() -> CGFloat {
-//        let inputAccessoryViewHeight = inputAccessoryView?.frame.height ?? 0
-        let inputContainerViewHeight = inputContainerView.frame.height
-        return max(0, inputContainerViewHeight + additionalBottomInset - automaticallyAddedBottomInset)
     }
 
     /// UIScrollView can automatically add safe area insets to its contentInset,
@@ -140,5 +106,9 @@ internal extension MessagesViewController {
     /// - Returns: The distance automatically added to contentInset.bottom, if any.
     private var automaticallyAddedBottomInset: CGFloat {
         return messagesCollectionView.adjustedContentInset.bottom - messagesCollectionView.contentInset.bottom
+    }
+
+    private var messageCollectionViewBottomInset: CGFloat {
+        return messagesCollectionView.contentInset.bottom
     }
 }
